Stripe & Firebase
I did way too many things in this commit all at once.

Firebase is in the application though it's not fully integrated.

The Donation Form for Stripe is working without user authentication, kind of like a guest checkout.

I'm going to need to clean this up in chunks but I'm too tired to do that right now.